AI can now reply to people who tag or reply to him.

+++ server/controllers/aiController.js
@@ -12,6 +12,9 @@ fs.readFile('./prompt.txt', 'utf8', function(err, data) {
 });
 
 exports.getAnswer = async function(res, req) {
+    chatLogs.replace('{botName}', req.body.botData.botName.split('#')[0]);
+    chatLogs.replace('{collectionName}', req.body.botData.collectionName);
+    chatLogs.replace('{mintDate}', req.body.botData.mintDate);
     chatLogs += `Human: ${req.body.text}\n`;
     const gptResponse = await openai.complete({
         engine: 'babbage',
@@ -56,7 +59,7 @@ exports.stopFarming = async function(res, req) {
 exports.updateBotSettings = async function(res, req) {
     const data = await levelFarms.findOne({ discordId: req.body.userToken, botName: req.body.botData.botName });
     if (data) {
-        await levelFarms.updateOne(data, { messageDelay: req.body.botData.messageDelay, channelId: req.body.botData.channelId });
+        await levelFarms.updateOne(data, { messageDelay: req.body.botData.messageDelay, channelId: req.body.botData.channelId, collectionName: req.body.botData.collectionName, mintDelay: req.body.botData.mintDelay });
         res.send({ state: 'success', message: 'Successfully updated settings' });
     } else {
         res.send({ state: 'error', message: 'Couldnt seem to find the bot' });
@@ -120,7 +123,9 @@ exports.startFarming = async function(res, req) {
                     messages: [],
                     botName: client.user.tag,
                     botAvatar: client.user.avatarURL(),
-                    botToken: req.body.token
+                    botToken: req.body.token,
+                    mintDate: req.body.mintDate,
+                    collectionName: req.body.collectionName
                 });
             } else {
                 await levelFarms.updateOne(checkIfBotExists, { running: true });
@@ -160,51 +165,57 @@ exports.startFarming = async function(res, req) {
                 }
                 if (message.author.bot) return;
                 if (message.channel.id != channelIdToCheck) return;
-                if (message.mentions.users.get(client.user.id).id === client.user.id) {
-                    currentlyChecking = true;
-                    const checkIfBotRunning = await levelFarms.findOne({ discordId: req.body.userToken, botName: client.user.tag });
-                    if (checkIfBotRunning) {
-                        if (checkIfBotRunning.running) {
-                            channelIdToCheck = checkIfBotRunning.channelId;
-                            (async() => {
-                                await sleep((10000 * Math.random()) + 1000);
-                                await axios({
-                                    method: 'post',
-                                    url: "https://beta.rudolphaio.com/api/askRudolph",
-                                    data: {
-                                        text: message.content, // This is the body part
-                                    }
-                                }).then(async function(response) {
-                                    response = response.data;
+                if (message.mentions.users.get(client.user.id)) {
+                    if (message.mentions.users.get(client.user.id).id === client.user.id) {
 
-                                    if (response == undefined || response == '') {
-                                        currentlyChecking = false;
-                                        return;
-                                    } else {
-                                        message.channel.send(`${response}`);
-                                    }
-                                    let data = checkIfBotRunning.messages;
-                                    data.push({ messageAuthor: message.author.tag, message: message.content, response: response });
-                                    if (data.length > 20) {
-                                        data.shift();
-                                    }
+                        currentlyChecking = true;
+                        const checkIfBotRunning = await levelFarms.findOne({ discordId: req.body.userToken, botName: client.user.tag });
+                        if (checkIfBotRunning) {
+                            if (checkIfBotRunning.running) {
+                                channelIdToCheck = checkIfBotRunning.channelId;
+                                (async() => {
+                                    await sleep((10000 * Math.random()) + 1000);
+                                    await axios({
+                                        method: 'post',
+                                        url: "https://beta.rudolphaio.com/api/askRudolph",
+                                        data: {
+                                            botData: checkIfBotRunning,
+                                            text: message.content, // This is the body part
+                                        }
+                                    }).then(async function(response) {
+                                        response = response.data;
 
-                                    await levelFarms.findOneAndUpdate({ discordId: req.body.userToken, botName: client.user.tag }, { messages: data });
-                                    minutesToAdd = checkIfBotRunning.messageDelay;
-                                    currentDate = new Date();
-                                    countDownDate = new Date(currentDate.getTime() + (minutesToAdd + 0.1) * 60000).getTime();
-                                    setTimeout(() => { currentlyChecking = false }, 1000);
-                                });
-                            })();
+                                        if (response == undefined || response == '') {
+                                            currentlyChecking = false;
+                                            return;
+                                        } else {
+                                            message.channel.send(`${response}`);
+                                        }
+                                        let data = checkIfBotRunning.messages;
+                                        data.push({ messageAuthor: message.author.tag, message: message.content, response: response });
+                                        if (data.length > 20) {
+                                            data.shift();
+                                        }
+
+                                        await levelFarms.findOneAndUpdate({ discordId: req.body.userToken, botName: client.user.tag }, { messages: data });
+                                        minutesToAdd = checkIfBotRunning.messageDelay;
+                                        currentDate = new Date();
+                                        countDownDate = new Date(currentDate.getTime() + (minutesToAdd + 0.1) * 60000).getTime();
+                                        setTimeout(() => { currentlyChecking = false }, 1000);
+                                    });
+                                })();
+                            } else {
+                                console.log('Shutting bot down...');
+                                clearInterval(x);
+                                client.destroy();
+                            }
                         } else {
                             console.log('Shutting bot down...');
                             clearInterval(x);
                             client.destroy();
                         }
                     } else {
-                        console.log('Shutting bot down...');
-                        clearInterval(x);
-                        client.destroy();
+                        return;
                     }
                 } else {
                     if (countDownDistance > 0 || currentlyChecking) return;
@@ -220,6 +231,7 @@ exports.startFarming = async function(res, req) {
                                     method: 'post',
                                     url: "https://beta.rudolphaio.com/api/askRudolph",
                                     data: {
+                                        botData: checkIfBotRunning,
                                         text: message.content, // This is the body part
                                     }
                                 }).then(async function(response) {
